+++ lfs_root/scripts/sync2lfs.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+set -e
+
+# unuse
+
+echo -e "--- start sync2lfs.sh ---\n\n"
+
+rsync -av /mnt/lfs_root ${LFS} && chown -R root:root ${LFSRoot}
+
+echo -e "+++ done sync2lfs.sh +++\n\n"